services/reddit_streaming_service/requirements.txt
requests == 2.32.3
praw == 7.7.1
tqdm == 4.66.5
pandas == 2.0.3
PyYAML == 6.0.1
kafka-python == 2.0.2
python-dotenv == 1.0.1